requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
db-dtypes==1.3.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
filelock==3.16.1
frozenlist==1.4.1
gcloud==0.18.3
google-api-core==2.21.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.26.0
google-cloud-bigquery-storage==2.26.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.28.0
itemadapter==0.9.0
itemloaders==1.3.2
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
lxml_html_clean==0.3.1
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
parsel==1.9.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
Protego==0.3.1
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.2
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==307
PyYAML==6.0.1
pyzmq==26.2.0
queuelib==1.7.0
readability-lxml==0.8.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
Scrapy==2.11.2
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tldextract==5.1.2
tornado==6.4.1
traitlets==5.14.3
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
w3lib==2.2.1
wcwidth==0.2.13
yarl==1.15.0
zope.interface==7.1.0